feat(ci): add CI
As we need to build three different targets I setup a matrix job which builds these targets in
parallel. This means I have to figure out how to add the corresponding tag, for the scratch image
I want to add no suffix, but for the other two I want to suffix the tag with the target name:
bullseye and alpine. For the resulting ternary expression I used a combination of `fromJSON` and
`format` to build the expression resulting in the correct values. The format step was necessary
as it was the only way I could inject the `matrix.target` in the array, as `fromJSON` expect a
string argument. I got the initial idea for the ternary expression from:
https://github.com/orgs/community/discussions/25725#discussioncomment-3248924