feat: add semantic-release
In order to automate releases I introduce semantic-release. This takes care of version bumping all
artifacts. As I want to publish our artifacts before I perform the actual release (as to avoid
having a release that is missing artifacts, due to publishing issues). I split the release into
two parts: firstly determine the version and using this version when publishing the artifacts.
Secondly create the actual release. Every action I'm using contains the actual permissions
necessary, to adhere to the principle of least privilege (I will also make sure we use the
default workflow permissions, that is, read-only access).

As semantic-release needs to push to a branch that is protected, we can't use the
GITHUB_TOKEN. Instead we create a new service user (founda-dev, a separate Github account)
which is allowed to bypass the branch protection. We create a Personal Access Token (PAT) for
this user and use that for semantic-release. I will use the same user in the `version` job so we
can at least verify it has push access to non-protected branches. *Note* in order for this to
work you have to disable credential caching during checkout, otherwise your token won't be used,
as the GITHUB_TOKEN is persisted.