fix: return dummy version in Dependabot PRs
When running this workflow in as the result of a Dependabot PR we don't have access to our
`FOUNDA_DEV_PAT`, but actually we don't even need to. As we don't publish artifacts when running
as the result of the Dependabot PR, there is also no need for a new version. So instead we will
always return `0.0.0` as the version.