fix: include Dependabot dependency updates in semantic-release
As Dependabot uses `chore(deps):` as the commit prefix we won't trigger a new release, but that
is not the behaviour we want. Dependency updates should trigger a new release, so we'll configure
semantic-release to map a commit of type `chore(deps):` to a patch release. *Note*, just to be
clear here: this means every `chore` commit we now be treated as the patch release.

See: https://mattbun.io/posts/semantic-release-dependabot/#approach-2---configure-semantic
-release-to-understand-dependabot-commits