build: fix setting wrong context
When moving the Dockerfile I by accident moved changed the `context` instead of the `file`
location, which caused the build failure as it never copied the `go.*` files. It did made me have
another look at the multi-arch builds, so let's call it a happy accident. Now the docker context
is correct again I will also add a `.dockerignore`.