fix: connect to the `postgres` database by default
I need to connect to a database before I can create one. Currently, as we don't specify a database
name, it tries to connect to a database named after the username, that's the default behaviour.
This is not what we want: this database is not guaranteed to exists and most likely never will.
So instead we will connect to the `postgres` database, this database seems like the safest choice
as far as picking a default database.